Refactor parking location
